refactor: resolve cyclic imports between scoring and state (#192)
Extract RoundAnalytics dataclass and _get_decade_label() into new
game/types.py module. Both scoring.py and state.py now import from
types.py instead of each other, eliminating all late/lazy imports
that worked around the circular dependency.

- New: game/types.py with shared types
- scoring.py: removed _get_decade_label def + late RoundAnalytics import
- state.py: removed RoundAnalytics class + late _get_decade_label import
- game/__init__.py: re-exports RoundAnalytics for backward compat
- All 129 tests pass
