+++ custom_components/beatify/game/__init__.py
@@ -2,5 +2,6 @@
 
 from .player import PlayerSession
 from .state import GamePhase, GameState
+from .types import RoundAnalytics
 
-__all__ = ["GamePhase", "GameState", "PlayerSession"]
+__all__ = ["GamePhase", "GameState", "PlayerSession", "RoundAnalytics"]
